Fix to support older git version
Older git versions (I use 2.1.4) don't support format specifier %cI, so
I we %ci here.

The end token of the here document must be proceeded with tab
characters, not spaces.

git filter-branch will only applied to the last $COMMITS versions.
This is a massive improvement on performance on large and / or complex
repositories.